Use https - now that it works.
Add a list of default libsets to check out when none are given on the
command line
  @def_libsets = qw{ GledCore Audio1 GTS Geom1 Numerica RootGeo Var1 };
and add this to help.

Check-out libsets into a libset/ directory and sym-link them to the
build-directory.

Do check-out of libsets and demos by using the 'update' method (instead of
'checkout') so that it is easier to use svn commands on all of them.